+++ lib/helpers/query/castUpdate.js
@@ -23,8 +23,13 @@ const utils = require('../../utils');
  * @param {Query} context passed to setters
  * @return {Boolean} true iff the update is non-empty
  */
-
 module.exports = function castUpdate(schema, obj, options, context, filter) {
+  console.log('hello Cast Update');
+  console.log(context);
+  console.log(context._mongooseOptions);
+  if (!('overwriteDiscriminatorKey' in context._mongooseOptions)) {
+    // do stuff
+  }
   if (obj == null) {
     return undefined;
   }
@@ -165,6 +170,9 @@ function castPipelineOperator(op, val) {
  */
 
 function walkUpdatePath(schema, obj, op, options, context, filter, pref) {
+  console.log('hello Path');
+  console.log(options);
+  console.log(context._mongooseOptions);
   const strict = options.strict;
   const prefix = pref ? pref + '.' : '';
   const keys = Object.keys(obj);
@@ -424,6 +432,9 @@ const overwriteOps = {
  */
 
 function castUpdateVal(schema, val, op, $conditional, context, path) {
+  console.log('hello Cast Update Val');
+  console.log(val);
+  console.log(context._mongooseOptions);
   if (!schema) {
     // non-existing schema path
     if (op in numberOps) {
