+++ test/model.discriminator.test.js
@@ -1686,4 +1686,34 @@ describe('model', function() {
 
     assert.deepEqual(nestedDocument.body.children[1].body.children[0].body.children[0].body.children, []);
   });
+
+  describe('Discriminator Key test', function() {
+    it('(gh-9015)', async function() {
+      const baseSchema = new Schema({}, { discriminatorKey: 'type' });
+          const baseModel = db.model('thing', baseSchema);
+          const aSchema = new Schema(
+            {
+              aThing: { type: Number },
+            },
+            { _id: false, id: false },
+          );
+          const aModel = baseModel.discriminator('A', aSchema);
+          const bSchema = new Schema(
+            {
+              bThing: { type: String },
+            },
+            { _id: false, id: false },
+          );
+          const bModel = baseModel.discriminator('B', bSchema);
+          // Model is created as a type A
+          const doc = await baseModel.create({ type: 'A', aThing: 1 });
+          const res = await baseModel.findByIdAndUpdate(
+            doc._id,
+            { type: 'A', bThing: 'one', aThing: '2' },
+            { runValidators: true, overwriteDiscriminatorKey: true },
+          );
+          assert.equal(res.type, 'B');
+    });
+  });
+
 });
